Commit 67: Añadida MapaView, accesible desde el HomeView_mobile utilizando el menú drawer, clicando el botón "MAPA". La vista MapaView mostrará los usuarios que se encuentren como máximo a 5  kilómetros de distancia del dispositivo.

+++ lib/FirestoreObjects/FbUsuario.dart
@@ -21,7 +21,7 @@ class FbUsuario {
     return FbUsuario(
         nombre: data?['nombre'] != null ? data!['nombre'] : "",
         edad: data?['edad'] != null ? data!['edad'] : 0,
-        altura: data?['altura'] != null ? data!['altura'] : 0,
+        altura: data?['altura'] != null ? data!['altura'].toDouble() : 0.0,
         geoloc:data?['geoloc'] != null ? data!['geoloc'] : const GeoPoint(0, 0)
     );
   }
